feature: addressing comments
