+++ Microsoft.Azure.Cosmos/src/direct/ISessionRetryOptions.cs
@@ -25,7 +25,6 @@
         Boolean RemoteRegionPreferred { get;}
 
 
-
     }
 }
